requirements.txt
requests
python-snappy
protobuf
